简化缩短版雪花算法生成id

+++ src/main/java/com/limyel/shutbb/dao/UsersTopicsMapper.java
@@ -3,13 +3,13 @@ package com.limyel.shutbb.dao;
 import com.limyel.shutbb.entity.UsersTopics;
 
 public interface UsersTopicsMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(UsersTopics record);
 
     int insertSelective(UsersTopics record);
 
-    UsersTopics selectByPrimaryKey(Integer id);
+    UsersTopics selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(UsersTopics record);
